:art: create a components MovieList.jsx with more data, improve request.js for better error management
