+++ client/src/services/request.js
@@ -2,9 +2,12 @@ import axios from "axios";
 
 function getMovies() {
   return axios
-    .get("http://localhost:3310/api/movies")
+    .get(`${import.meta.env.VITE_API_URL}/api/movies`)
     .then((response) => response.data)
-    .catch((error) => console.error(error));
+    .catch((error) => {
+      console.error(error);
+      return [];
+    });
 }
 
 export default getMovies;
